// src/db/schema.ts
export const clientSelfiesTable = pgTable("pdc_selfies", {
  selfieId: text("selfie_id").notNull().primaryKey(), // selfie id
  selfieUrl: text("selfie_url").notNull(), // selfie url
  selfieThumbnail: text("selfie_thumbnail").notNull(), // selfie thumbnail
  shiftX: real("shift_x"), // selfie shift x
  shiftY: real("shift_y"), // selfie shift y
  zoom: real("zoom"), // selfie zoom
  width: real("width"), // selfie width
  height: real("height"), // selfie height
  createdAt: timestamp("created_at").notNull().defaultNow(), // date of creation
});

export const clientTable = pgTable("pdc_client", {
  clientId: text("client_id").notNull().primaryKey(), // client id
  createdAt: timestamp("created_at").notNull().defaultNow(), // date of creation
  phone: text("phone").notNull(), // client phone
  selfieId: text("selfie_id").references(() => clientSelfiesTable.selfieId), // ref for selfie
  email: text("email"), // optional email
  fullName: text("full_name"), // optional name
});